Halbrügge, M. & Russwinkel, N. (2016). The Sum of Two Models: How a Composite Model Explains Unexpected User Behavior in a Dual-Task Scenario. In Proc. ICCM 2016
Abstract: Maintaining cognitive control while pursuing several tasks at the same time is hard, especially when the current problem states of these tasks need to be represented in memory. We are investigating the mutual influence of a self-paced and a reactive task with regard to completion time and error rates. Against initial expectations, the interruptions from the reactive task did not lead to more errors in the self-paced task, but only prolonged the completion time. Our understanding of this result is guided by a combined version of two previously published cognitive models of the individual tasks. The combined model reproduces the empirical findings concerning error rates and task completion times, but not an unexpected change in the error pattern. These results feed back into our theoretical understanding of cognitive control during sequential action.